fix: update Lambda Python version 3.8 -> 3.13 (#3338)
Python 3.8 and 3.9 are end of life, upgrade to a recent supported Python
version.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

+++ packages/amplify-e2e-core/src/utils/index.ts
@@ -29,6 +29,10 @@ config();
 export const TEST_PROFILE_NAME = 'amplify-integ-test-user';
 
 export function deleteProjectDir(root: string) {
+  if (process.env.SKIP_DELETE) {
+    console.warn(`🌋 Did not delete project dir: ${root}`);
+    return;
+  }
   rimraf.sync(root);
 }
